:spearkles: Add new Method Deposit
