+++ Parametros.go
@@ -1,11 +1,11 @@
-package estudandoGo
+package main
 
-import (
-	"fmt"
-)
+import "fmt"
 
+
+// explicando parametro
 func semParametro() string {
-	return "exempl de funcao sem paramentro!"
+	return "exemplo de funcao sem paramentro!"
 }
 
 func umParametro(texto string) string {
@@ -16,10 +16,9 @@ func doisParametros (texto string, numero int) (string, int) {
 	return texto, numero
 }
 
-func parametros()  {
+func parametro ()  {
 	fmt.Println(semParametro())
 	fmt.Println(umParametro("um parametro"))
 	fmt.Println(doisParametros("dois parametros: eessa string e o numero", 10))
-
 
 }
